Fallback on Moo::Role/Role::Tiny role when applying
This simplistic commit checks if the parameters are empty ({}) and if they are,
when a role package does not have apply(), fallback to applying using
Moo::Role or Role::Tiny.

This commit is basically just a proof of concept to make classes that use
MooX::Role::Parameterized::With work with Role::Tiny and Moo::Role roles.

One caveat is that the required methods must be declared before including the
role, e.g.:

    package Role1;
    use Role::Tiny;
    requires 'meth1';

    package Class1;
    sub meth1 { ... }
    use MooX::Role::Parameterized::With 'Role1' => {};

A solution for that is for MooX::Role::Parameterized::With to declare 'with'. A
syntax that is compatible with Moose to support non-parametric and parametric
including would be nice, e.g.:

    with 'Role1', 'Role2' => {param=>'for',role=>'Role2'}, ...;